requirements.txt
Flask==3.0.0
anthropic==0.40.0
python-dotenv==1.0.0
Flask-Session==0.5.0
flask-talisman==1.1.0
pyOpenSSL==23.2.0  # For SSL support
